requirements.txt
python-okx~=0.3.3
SQLAlchemy~=2.0.36
python-dotenv~=1.0.1